PCB processes, inputs and outputs are now built into the pcb. Each pcb should have its own class. Instructions have been removed. the new implimentation should make it easy to make completely custom pcb components.

+++ src/main/java/net/cubic/dronecraft_2/dronecraft_2/data/PCB/PCBVarTypes.java
@@ -28,7 +28,7 @@ public class PCBVarTypes {
 
     @SubscribeEvent
     public static void onPCBRegistry(final RegistryEvent.Register<VarType> event) {
-        NUMBER = new VarType(new RGBA(255,255,0),Float.class).setRegistryName("number");//TODO add custom wire items that can be used to craft them
+        NUMBER = new VarType(new RGBA(255,255,0),Double.class).setRegistryName("number");//TODO add custom wire items that can be used to craft them
         BOOLEAN = new VarType(new RGBA(240,10,10),Boolean.class).setRegistryName("boolean");
         XYZ = new VarType(new RGBA(255,40,255), Vector3f.class).setRegistryName("xyz");
